+++ primefaces/src/test/java/org/primefaces/util/CurrencyValidatorTest.java
@@ -30,7 +30,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
-import java.text.DecimalFormatSymbols;
+import java.text.DecimalFormat;
 import java.util.Locale;
 
 /**
@@ -42,15 +42,17 @@ public class CurrencyValidatorTest {
     private static String ukPound;
     private static String brazilReal;
     private static String customCurrencySymbol;
-    private static Locale brazil = new Locale("pt", "BR");
-    private static Locale custom = new Locale("es", "US");
+    private static DecimalFormat uk = (DecimalFormat) DecimalFormat.getCurrencyInstance(Locale.UK);
+    private static DecimalFormat us = (DecimalFormat) DecimalFormat.getCurrencyInstance(Locale.US);
+    private static DecimalFormat brazil = (DecimalFormat) DecimalFormat.getCurrencyInstance(new Locale("pt", "BR"));
+    private static DecimalFormat custom = (DecimalFormat) DecimalFormat.getCurrencyInstance(new Locale("es", "US"));
 
     @BeforeAll
     protected static void setUp() throws Exception {
-        usDollar = (new DecimalFormatSymbols(Locale.US)).getCurrencySymbol();
-        ukPound = (new DecimalFormatSymbols(Locale.UK)).getCurrencySymbol();
-        brazilReal = (new DecimalFormatSymbols(brazil)).getCurrencySymbol();
-        customCurrencySymbol = (new DecimalFormatSymbols(custom)).getCurrencySymbol();
+        usDollar = us.getDecimalFormatSymbols().getCurrencySymbol();
+        ukPound = uk.getDecimalFormatSymbols().getCurrencySymbol();
+        brazilReal = brazil.getDecimalFormatSymbols().getCurrencySymbol();
+        customCurrencySymbol = custom.getDecimalFormatSymbols().getCurrencySymbol();
     }
 
     private static int getVersion() {
@@ -78,23 +80,23 @@ public class CurrencyValidatorTest {
         BigDecimal noDecimal = new BigDecimal("1234.00");
         BigDecimal oneDecimal = new BigDecimal("1234.50");
 
-        assertEquals(expected, validator.validate(ukPound + "1,234.56", Locale.UK), "UK locale");
-        assertEquals(negative, validator.validate("-" + ukPound + "1,234.56", Locale.UK), "UK negative");
-        assertEquals(noDecimal, validator.validate(ukPound + "1,234", Locale.UK), "UK no decimal");
-        assertEquals(oneDecimal, validator.validate(ukPound + "1,234.5", Locale.UK), "UK 1 decimal");
-        assertEquals(expected, validator.validate(ukPound + "1,234.567", Locale.UK), "UK 3 decimal");
-        assertEquals(expected, validator.validate("1,234.56", Locale.UK), "UK no symbol");
+        assertEquals(expected, validator.validate(ukPound + "1,234.56", uk), "UK locale");
+        assertEquals(negative, validator.validate("-" + ukPound + "1,234.56", uk), "UK negative");
+        assertEquals(noDecimal, validator.validate(ukPound + "1,234", uk), "UK no decimal");
+        assertEquals(oneDecimal, validator.validate(ukPound + "1,234.5", uk), "UK 1 decimal");
+        assertEquals(expected, validator.validate(ukPound + "1,234.567", uk), "UK 3 decimal");
+        assertEquals(expected, validator.validate("1,234.56", uk), "UK no symbol");
 
-        assertEquals(expected, validator.validate(usDollar + "1,234.56", Locale.US), "US locale");
-        assertEquals(noDecimal, validator.validate(usDollar + "1,234", Locale.US), "US no decimal");
-        assertEquals(oneDecimal, validator.validate(usDollar + "1,234.5", Locale.US), "US 1 decimal");
-        assertEquals(expected, validator.validate(usDollar + "1,234.567", Locale.US), "US 3 decimal");
-        assertEquals(expected, validator.validate("1,234.56", Locale.US), "US no symbol");
+        assertEquals(expected, validator.validate(usDollar + "1,234.56", us), "US locale");
+        assertEquals(noDecimal, validator.validate(usDollar + "1,234", us), "US no decimal");
+        assertEquals(oneDecimal, validator.validate(usDollar + "1,234.5", us), "US 1 decimal");
+        assertEquals(expected, validator.validate(usDollar + "1,234.567", us), "US 3 decimal");
+        assertEquals(expected, validator.validate("1,234.56", us), "US no symbol");
         if (getVersion() > 8) {
-            assertEquals(negative, validator.validate("-" + usDollar + "1,234.56", Locale.US), "US negative");
+            assertEquals(negative, validator.validate("-" + usDollar + "1,234.56", us), "US negative");
         }
         else {
-            assertEquals(negative, validator.validate("(" + usDollar + "1,234.56)", Locale.US), "US negative");
+            assertEquals(negative, validator.validate("(" + usDollar + "1,234.56)", us), "US negative");
         }
     }
 
@@ -107,15 +109,15 @@ public class CurrencyValidatorTest {
         BigDecimal expected = new BigDecimal("1234.00");
         BigDecimal negative = new BigDecimal("-1234.00");
 
-        assertEquals(expected, validator.validate(ukPound + "1,234", Locale.UK), "UK locale");
-        assertEquals(negative, validator.validate("-" + ukPound + "1,234", Locale.UK), "UK negative");
+        assertEquals(expected, validator.validate(ukPound + "1,234", uk), "UK locale");
+        assertEquals(negative, validator.validate("-" + ukPound + "1,234", uk), "UK negative");
 
-        assertEquals(expected, validator.validate(usDollar + "1,234", Locale.US), "US locale");
+        assertEquals(expected, validator.validate(usDollar + "1,234", us), "US locale");
         if (getVersion() > 8) {
-            assertEquals(negative, validator.validate("-" + usDollar + "1,234", Locale.US), "US negative");
+            assertEquals(negative, validator.validate("-" + usDollar + "1,234", us), "US negative");
         }
         else {
-            assertEquals(negative, validator.validate("(" + usDollar + "1,234)", Locale.US), "US negative");
+            assertEquals(negative, validator.validate("(" + usDollar + "1,234)", us), "US negative");
         }
     }
 
@@ -127,8 +129,8 @@ public class CurrencyValidatorTest {
         CurrencyValidator validator = CurrencyValidator.getInstance();
         BigDecimal expected = new BigDecimal("20.00");
 
-        assertEquals(expected, validator.validate("¤20.00", Locale.US), "US locale");
-        assertEquals(expected, validator.validate("¤20.00", Locale.UK), "UK locale");
+        assertEquals(expected, validator.validate("¤20.00", us), "US locale");
+        assertEquals(expected, validator.validate("¤20.00", uk), "UK locale");
     }
 
     /**
@@ -138,8 +140,8 @@ public class CurrencyValidatorTest {
     public void testExcelPatterns() {
         CurrencyValidator validator = CurrencyValidator.getInstance();
 
-        assertEquals("\"" + usDollar + "\"" + "#,##0.00", validator.getExcelPattern(Locale.US), "US");
-        assertEquals("\"" + ukPound + "\"" +  "#,##0.00", validator.getExcelPattern(Locale.UK), "UK");
+        assertEquals("\"" + usDollar + "\"" + "#,##0.00", validator.getExcelPattern(us), "US");
+        assertEquals("\"" + ukPound + "\"" +  "#,##0.00", validator.getExcelPattern(uk), "UK");
         assertEquals("\"" + customCurrencySymbol + "\"#,##0.00", validator.getExcelPattern(custom), "Custom");
     }
 
@@ -150,7 +152,7 @@ public class CurrencyValidatorTest {
     public void testWeirdPatternIsNotParsed() {
         CurrencyValidator validator = CurrencyValidator.getInstance();
 
-        Number result = validator.validate("74E12341", Locale.UK);
+        Number result = validator.validate("74E12341", uk);
         assertNull(result);
     }